pyproject.toml
[project]
name = "rate-limiter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "bcrypt<4",
    "fastapi[standard]>=0.123.9",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.44",
]
[tool.ruff.lint]
ignore = ["INP001"]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "pytest>=9.0.2",
]
[tool.pytest.ini_options]
pythonpath = [
  "."
]
